refactor(web): extract kilo schema extras into their own file
Isolates the Kilo-specific additions from the merge/transport logic so
adding a new schema key is a single-file change with a documented
step-by-step in extras.ts.
